Merge in MB/mb-backend from MB-504-add_inn to develop

* commit '3b8e1780e8a2b6030f14086f20234055221a9f2d':
  [MB-504] добавить Inn

+++ api/src/main/java/pro/mbroker/api/dto/BorrowerEmployerDto.java
@@ -18,7 +18,7 @@ public class BorrowerEmployerDto {
 
     private String name;
 
-    private Integer inn;
+    private Long inn;
 
     private Branch branch;
